feat: server side error handling

+++ src/main/java/com/skecorp/studentmgmt/student/StudentService.java
@@ -1,5 +1,7 @@
 package com.skecorp.studentmgmt.student;
 
+import com.skecorp.studentmgmt.student.exception.BadRequestException;
+import com.skecorp.studentmgmt.student.exception.StudentNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -15,10 +17,23 @@ public class StudentService {
     }
 
     public void addStudent(Student student) {
+        // check if email is taken
+        var studentExists = studentRepository.selectExistsEmail(student.getEmail());
+        if (studentExists){
+            throw new BadRequestException(String.format(
+                    "Student with email %s already exists", student.getEmail()));
+        }
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId) {
+        // check if student exists
+         var studentExists = studentRepository.findById(studentId);
+         if(studentExists == null) {
+             throw new StudentNotFoundException(String.format(
+                     "student with id: %s does not exist",
+                     studentId));
+         }
         studentRepository.deleteById(studentId);
     }
 }
